feat(model.event.EventType) : EventClassification 필드 추가
